Move file to shared
