+++ src/screens/authentication/SignUpScreen.tsx
@@ -15,7 +15,7 @@ import { SafeAreaView } from 'react-native-safe-area-context'
 import { FormViolation } from '../../core/entities/DetailedProfile'
 import { Colors, Spacing, Styles, Typography } from '../../styles'
 import i18n from '../../utils/i18n'
-import BirthdayPicker from '../personalInformation/BirthdayPicker'
+import BirthdayPicker from '../shared/BirthdayPicker'
 import LabelInputContainer from '../personalInformation/LabelInputContainer'
 import LocationPicker from '../personalInformation/LocationPicker'
 import PhoneNumberInput from '../personalInformation/PhoneNumberInput'
