+++ src/screens/personalInformation/PersonalInformationScreenContent.tsx
@@ -6,7 +6,7 @@ import KeyboardOffsetView from '../shared/KeyboardOffsetView'
 import LabelInputContainer from './LabelInputContainer'
 import LabelTextInput from './LabelTextInput'
 import GenderPicker from './GenderPicker'
-import BirthdayPicker from './BirthdayPicker'
+import BirthdayPicker from '../shared/BirthdayPicker'
 import CountryPicker from 'react-native-country-picker-modal'
 import LocationPicker from './LocationPicker'
 import { FormViolation } from '../../core/entities/DetailedProfile'
